requirements.txt
asgiref==3.4.0
autopep8==1.5.7
chardet==4.0.0
charset-normalizer==2.0.4
Django==3.2.5
djangorestframework==3.12.4
idna==2.10
Jinja2==3.0.1
MarkupSafe==2.0.1
packaging==20.9
pycodestyle==2.7.0
pyparsing==2.4.7
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.6
webencodings==0.5.1
zipp==3.4.1
